simplify test node
