+++ tests/web_node/mod.rs
@@ -1,6 +1,39 @@
 #![allow(dead_code)]
 #![allow(unused_imports)]
 
-pub mod provider;
+mod block_store;
+mod event_stream;
+pub mod keystore;
 
-pub use provider::{ProviderImpl, keystore};
+use anyhow::Result;
+use blockstore::InMemoryBlockstore;
+use credibil_dwn::provider::{
+    DataStore, DidDocument, DidResolver, EventLog, MessageStore, Provider, TaskStore,
+};
+
+#[derive(Clone)]
+pub struct ProviderImpl {
+    blockstore: InMemoryBlockstore<64>,
+    pub nats_client: async_nats::Client,
+}
+
+impl ProviderImpl {
+    pub async fn new() -> Result<Self> {
+        Ok(Self {
+            blockstore: InMemoryBlockstore::<64>::new(),
+            nats_client: async_nats::connect("demo.nats.io").await?,
+        })
+    }
+}
+
+impl Provider for ProviderImpl {}
+impl MessageStore for ProviderImpl {}
+impl DataStore for ProviderImpl {}
+impl EventLog for ProviderImpl {}
+impl TaskStore for ProviderImpl {}
+
+impl DidResolver for ProviderImpl {
+    async fn resolve(&self, _url: &str) -> Result<DidDocument> {
+        unimplemented!("DidResolver::resolve")
+    }
+}
